iOS: Improve import copy safety and structure
